Completes all required tests
